+++ Examples/Resources/SqlAGReplica/2-RemoveAvailabilityGroupReplica.ps1
@@ -90,8 +90,8 @@ Configuration Example
                 Ensure                        = 'Absent'
                 Name                          = $Node.NodeName
                 AvailabilityGroupName         = $Node.AvailabilityGroupName
-                SQLServer                     = $Node.NodeName
-                SQLInstanceName               = $Node.SQLInstanceName
+                ServerName                    = $Node.NodeName
+                InstanceName                  = $Node.SQLInstanceName
                 PrimaryReplicaSQLServer       = ( $AllNodes | Where-Object { $_.Role -eq 'PrimaryReplica' } ).NodeName
                 PrimaryReplicaSQLInstanceName = ( $AllNodes | Where-Object { $_.Role -eq 'PrimaryReplica' } ).SQLInstanceName
             }