+++ CHANGELOG.md
@@ -51,6 +51,8 @@
     the resource will only determine if a change is needed if the target node
     is the active host of the SQL Server instance ([issue #869](https://github.com/PowerShell/xSQLServer/issues/869)).
 - Changes to xSQLServerAlwaysOnAvailabilityGroupReplica
+  - BREAKING CHANGE: Parameters SQLServer and SQLInstanceName has been renamed
+    to ServerName and InstanceName respectivly ([issue #308](https://github.com/PowerShell/xSQLServer/issues/)).
   - Made the resource cluster aware. When ProcessOnlyOnActiveNode is specified,
     the resource will only determine if a change is needed if the target node is
     the active host of the SQL Server instance ([issue #870](https://github.com/PowerShell/xSQLServer/issues/870)).